fix goroutine leak in event forwarder / fanout
The fanout logic assumed that the subscriber context would be cancelled fairly
quick, as it was meant to be used in RPC request processing. But the event
forwarder used a long-lived context that wouldn't cancel until it hits an error
that requires a restart.

The forwarder now uses a context that gets cancelled after each eventlog poll
run.