let postgres ensure that only one replicator is running
As we're using separate connections for the tx lock and and the replication they
could get out of sync.

This changes the logic so that we instead rely on postgres telling us whether a
replication is active or not. We do this by acting either on the replication
slot being active, or if `START_REPLICATION SLOT` returns the error code
55006 (object_in_use).

This will prevent the many...

```
failed to start replication: ERROR: replication slot "eventlogslot" is active for PID 89499 (SQLSTATE 55006)
```

...log entries that we have been getting.