change how we do restores, emit all events
As the eventlog drives the archiving we have to create events for all document
versions and statuses. Previously we just emitted the final document version and
status for each named status.

Also changed the archived reference counting so that all events for the document
must be archived before we proceed. Less special-casing, more robust.