Feature/archive eventlog (#463)
* handle documents that have been deleted before they can be archived

* update linter

* change how we do restores, emit all events

As the eventlog drives the archiving we have to create events for all document
versions and statuses. Previously we just emitted the final document version and
status for each named status.

Also changed the archived reference counting so that all events for the document
must be archived before we proceed. Less special-casing, more robust.

* consistenly set original creator as creator on restore

add restoration metadata to statuses as well as document versions

* cover more behaviours around restore in tests

* make history test data temporally stable

* use the golden file test helper

* adding nonce to eventlog and get meta, test updates

Adding to the API response caused a lot of changes i tests. Primarily because I
took the opportunity to switch to the golden file helper as well. That caused a
lot of changes in testdata, but mainly in formatting changes and attribute
naming that is consistent with protojson.

* verify document signature against its event when restoring.