use delete manifest as source of truth for restores (#494)
The restore process previously listed the objects in S3 and then verified that
they matched the objects that we could expect based on the manifest. We now
instead use the manifest to fetch the objects we expect to be there, resulting
in a much leaner codebase and increased robustness, as we won't be failing if
unknown/unexpected objects exist in the bucket.